+++ src/main/java/com/minhcraft/beyondbetatweaks/mixin/entity/AbstractHorseMixin.java
@@ -1,10 +1,16 @@
 package com.minhcraft.beyondbetatweaks.mixin.entity;
 
 import com.minhcraft.beyondbetatweaks.config.ModConfig;
+import net.minecraft.world.entity.AgeableMob;
+import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.animal.horse.AbstractHorse;
+import net.minecraft.world.entity.animal.horse.Donkey;
+import net.minecraft.world.entity.animal.horse.Horse;
+import net.minecraft.world.entity.animal.horse.Mule;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import java.util.function.DoubleSupplier;
@@ -40,5 +46,29 @@ public abstract class AbstractHorseMixin {
     )
     private static void beyond_beta_tweaks$overrideHorseMaxHealth(IntUnaryOperator operator, CallbackInfoReturnable<Float> cir) {
         cir.setReturnValue(20.0F);
+    }
+
+    @Inject(
+            method = "setOffspringAttributes",
+            at = @At("HEAD"),
+            cancellable = true
+    )
+    private void beyond_beta_tweaks$hardcodeBredHorseStats(AgeableMob parent, AbstractHorse child, CallbackInfo ci) {
+        if ( child instanceof Donkey) {
+            child.getAttribute(Attributes.MAX_HEALTH).setBaseValue(20.0);
+            child.getAttribute(Attributes.JUMP_STRENGTH).setBaseValue(0.5);
+            child.getAttribute(Attributes.MOVEMENT_SPEED).setBaseValue(ModConfig.donkeyMovementSpeed);
+            ci.cancel();
+        } else if (child instanceof Horse) {
+            child.getAttribute(Attributes.MAX_HEALTH).setBaseValue(20.0);
+            child.getAttribute(Attributes.JUMP_STRENGTH).setBaseValue(0.6);
+            child.getAttribute(Attributes.MOVEMENT_SPEED).setBaseValue(ModConfig.horseMovementSpeed);
+            ci.cancel();
+        } else if ( child instanceof Mule) {
+            child.getAttribute(Attributes.MAX_HEALTH).setBaseValue(20.0);
+            child.getAttribute(Attributes.JUMP_STRENGTH).setBaseValue(0.6);
+            child.getAttribute(Attributes.MOVEMENT_SPEED).setBaseValue(ModConfig.donkeyMovementSpeed);
+            ci.cancel();
+        }
     }
 }